reworked node::buffer*
buffer_view is now the base-class of buffer and hashed_buffer_view,
buffer will be the base-class of hashed_buffer. hashed_buffer_view acts
as a middleman between hashed_buffer and buffer_view signaling that a
hash value might be present. Furthermore literal_buffer_view is the
return value of ""_view and signals to other classes that the backing
data is static and thus node::weak can safely be assumed (without
requiring the usual "explicit" instantiation).